ci: run tests and golangci-lint on the core backend
publish-core.yml built and pushed the image and nothing else, so no push or
pull request ever ran go test or a linter — despite the suite passing cleanly
under -race and golangci-lint being installed locally with no config to use.
The unchecked CreateIndexes errors fixed two commits ago sat in the tree
because nothing was looking.

Add core/.golangci.yml with 13 linters and a CI workflow running build,
go mod tidy -diff, go test -race and lint. errcheck is the reason the config
exists: it is the linter that catches the dropped-error class go vet misses.

The enabled set is green as of this commit, which took a handful of fixes:
three redundant int conversions, three dead declarations (boolPtr, boolDeref,
and a vestigial `_ = key{}` keeping an unused type alive), one redundant
Go 1.22 loop-variable copy, gofmt on three files, and seven unchecked Close
calls in error paths made explicit discards. A gate that starts red gets
switched off, so the config keeps nothing failing.

nilerr and contextcheck are documented in the config as deliberately
excluded rather than silently omitted. nilerr's 23 hits are almost all the
intentional "row was deleted, render null" pattern; it should go on once a
repository.ErrNotFound sentinel lands, at which point a Mongo timeout stops
being indistinguishable from a missing row and the remaining hits become the
bugs the linter is for.

+++ core/pkg/resolver/wiki_document_visit_resolver.go
@@ -140,7 +140,7 @@ func (r *wikiDocumentVisitResolver) WikiDocumentHistory(ctx context.Context, ope
 	if offset != nil && *offset > 0 {
 		off = int64(*offset)
 	}
-	lim := int64(MaxWikiVisitHistory)
+	lim := MaxWikiVisitHistory
 	if limit != nil && *limit > 0 {
 		lim = int64(*limit)
 		if lim > MaxWikiVisitHistory {
